+++ panda/plugins/taint2/taint_ops.h
@@ -78,7 +78,8 @@ void taint_mix_compute(Shad *shad, uint64_t dest, uint64_t dest_size,
 //for mul or fmul. can do parallel or mixed or no prop depending on vals and their taints
 void taint_mul_compute(Shad *shad, uint64_t dest, uint64_t dest_size,
                        uint64_t src1, uint64_t src2, uint64_t src_size,
-                       llvm::Instruction *inst, uint64_t arg1, uint64_t arg2);
+                       llvm::Instruction *inst, unsigned __int128 arg1,
+                       unsigned __int128 arg2);
 
 // Clear taint.
 void taint_delete(Shad *shad, uint64_t dest, uint64_t size);
